+++ projects/ccd-case-ui-toolkit/src/lib/components/banners/notification-banner/domain/notification-banner-config.model.ts
@@ -1,13 +1,14 @@
+import { Observable } from 'rxjs';
 import { NotificationBannerType } from '../enums/notification-banner-type.enum';
 
 export interface NotificationBannerConfig {
   bannerType: NotificationBannerType;
-  headingText: string;
-  description: string;
+  headingText: Observable<string>;
+  description: Observable<string>;
   showLink: boolean;
-  linkUrl?: string;
-  linkText?: string;
+  linkUrl?: Observable<string>;
+  linkText?: Observable<string>;
   triggerOutputEvent?: boolean;
-  triggerOutputEventText?: string;
+  triggerOutputEventText?: Observable<string>;
   headerClass: string;
 }