+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/case-editor/case-edit/case-edit.component.spec.ts
@@ -866,7 +866,7 @@ describe('CaseEditComponent', () => {
 
         it('should check page is not refreshed', () => {
           mockSessionStorageService.getItem.and.returnValue(component.initialUrl = null);
-          mockSessionStorageService.getItem.and.returnValue(component.isPageRefreshed = false);
+          mockSessionStorageService.getItem.and.returnValue((component.isPageRefreshed = false).toString());
 
           fixture.detectChanges();
           expect(component.checkPageRefresh()).toBe(false);
@@ -874,7 +874,7 @@ describe('CaseEditComponent', () => {
 
         it('should check page is refreshed', () => {
           mockSessionStorageService.getItem.and.returnValue(component.initialUrl = 'test');
-          mockSessionStorageService.getItem.and.returnValue(component.isPageRefreshed = true);
+          mockSessionStorageService.getItem.and.returnValue((component.isPageRefreshed = true).toString());
 
           fixture.detectChanges();
           expect(component.checkPageRefresh()).toBe(true);