+++ projects/ccd-case-ui-toolkit/src/lib/shared/domain/work-allocation/task-response.model.ts
@@ -1,5 +1,5 @@
 import { Task } from './Task';
 
-export interface TaskRespone {
+export interface TaskResponse {
   task: Task;
 }